scripts/SWDTokenSetNotifications/requirements.txt
discord.py
pexpect
python-dotenv
requests
web3
python-telegram-bot